+++ CHANGELOG.md
@@ -91,6 +91,7 @@
 - PR #2794 Fix async race in NVCategory::get_value and get_value_bounds
 - PR #2795 Fix java build/cast error
 - PR #2496 Fix improper merge of two dataframes when names differ
+- PR #2765 Fix Java inequality comparisons for string category
 - PR #2818 Fix java join API to use new C++ join API